// src/api/model/aPIError.ts
/**
 * Pizzeria Fabulosa API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface APIError { 
    id?: number;
    cause: string;
    message: string;
    origin: string;
    path: string;
    logged: boolean;
    fatal: boolean;
    createdOn: string;
}