// src/api/model/responseDTO.ts
/**
 * Pizzeria Fabulosa API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { APIError } from './aPIError';


export interface ResponseDTO { 
    apiError: APIError;
}